+++ OpenAgua/__init__.py
@@ -4,7 +4,8 @@ import os
 from flask import Flask
 
 # import blueprints
-from .projects_manager import projects
+from .user_projects import user_projects
+from .user_home import user_home
 from .network_editor import net_editor
 from .model_dashboard import model_dashboard
 
@@ -16,6 +17,7 @@ app.config.from_pyfile('config.py')
 import OpenAgua.views
 
 # register blueprints
-app.register_blueprint(projects, url_prefix='')
+app.register_blueprint(user_home, url_prefix='')
+app.register_blueprint(user_projects, url_prefix='')
 app.register_blueprint(net_editor, url_prefix='')
 app.register_blueprint(model_dashboard, url_prefix='')
