+++ OpenAgua/user_home/__init__.py
@@ -1,9 +1,9 @@
 from flask import Blueprint
 
-projects = Blueprint('projects',
+user_home = Blueprint('user_home',
                        __name__,
                        template_folder='templates',
                        static_folder='static',
-                       static_url_path='/projects/static')
+                       static_url_path='/user_home/static')
 
 from . import views
