+++ OpenAgua/user_projects/__init__.py
@@ -1,0 +1,9 @@
+from flask import Blueprint
+
+user_projects = Blueprint('user_projects',
+                       __name__,
+                       template_folder='templates',
+                       static_folder='static',
+                       static_url_path='/user_projects/static')
+
+from . import views
